requirements.txt
rapidfuzz==3.6.1
pyarabic==0.6.15
pandas==2.2.0
openpyxl==3.1.2
jinja2==3.1.3
tqdm==4.66.1
python-dotenv==1.0.1